// src/app/lib/generated/models/ChatbotType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ChatbotType = {
    Girls: 'girls',
    Anime: 'anime',
    Boys: 'boys',
    Internal: 'internal',
    Fantasy: 'fantasy'
} as const;
export type ChatbotType = typeof ChatbotType[keyof typeof ChatbotType];


export function instanceOfChatbotType(value: any): boolean {
    for (const key in ChatbotType) {
        if (Object.prototype.hasOwnProperty.call(ChatbotType, key)) {
            if (ChatbotType[key as keyof typeof ChatbotType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ChatbotTypeFromJSON(json: any): ChatbotType {
    return ChatbotTypeFromJSONTyped(json, false);
}

export function ChatbotTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ChatbotType {
    return json as ChatbotType;
}

export function ChatbotTypeToJSON(value?: ChatbotType | null): any {
    return value as any;
}

export function ChatbotTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ChatbotType {
    return value as ChatbotType;
}