// src/app/lib/generated/models/BreastSize.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const BreastSize = {
    Small: 'small',
    Medium: 'medium',
    Large: 'large',
    Huge: 'huge'
} as const;
export type BreastSize = typeof BreastSize[keyof typeof BreastSize];


export function instanceOfBreastSize(value: any): boolean {
    for (const key in BreastSize) {
        if (Object.prototype.hasOwnProperty.call(BreastSize, key)) {
            if (BreastSize[key as keyof typeof BreastSize] === value) {
                return true;
            }
        }
    }
    return false;
}

export function BreastSizeFromJSON(json: any): BreastSize {
    return BreastSizeFromJSONTyped(json, false);
}

export function BreastSizeFromJSONTyped(json: any, ignoreDiscriminator: boolean): BreastSize {
    return json as BreastSize;
}

export function BreastSizeToJSON(value?: BreastSize | null): any {
    return value as any;
}

export function BreastSizeToJSONTyped(value: any, ignoreDiscriminator: boolean): BreastSize {
    return value as BreastSize;
}