// src/app/lib/generated/models/ButtSize.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ButtSize = {
    Small: 'small',
    Medium: 'medium',
    Large: 'large',
    Huge: 'huge'
} as const;
export type ButtSize = typeof ButtSize[keyof typeof ButtSize];


export function instanceOfButtSize(value: any): boolean {
    for (const key in ButtSize) {
        if (Object.prototype.hasOwnProperty.call(ButtSize, key)) {
            if (ButtSize[key as keyof typeof ButtSize] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ButtSizeFromJSON(json: any): ButtSize {
    return ButtSizeFromJSONTyped(json, false);
}

export function ButtSizeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ButtSize {
    return json as ButtSize;
}

export function ButtSizeToJSON(value?: ButtSize | null): any {
    return value as any;
}

export function ButtSizeToJSONTyped(value: any, ignoreDiscriminator: boolean): ButtSize {
    return value as ButtSize;
}