// src/app/lib/generated/models/SpecialFeatures.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI Babe Public API
 * Lorem ipsum
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const SpecialFeatures = {
    Tattoos: 'tattoos',
    PubicHair: 'pubic_hair',
    Piercing: 'piercing',
    Freckles: 'freckles',
    Pregnant: 'pregnant'
} as const;
export type SpecialFeatures = typeof SpecialFeatures[keyof typeof SpecialFeatures];


export function instanceOfSpecialFeatures(value: any): boolean {
    for (const key in SpecialFeatures) {
        if (Object.prototype.hasOwnProperty.call(SpecialFeatures, key)) {
            if (SpecialFeatures[key as keyof typeof SpecialFeatures] === value) {
                return true;
            }
        }
    }
    return false;
}

export function SpecialFeaturesFromJSON(json: any): SpecialFeatures {
    return SpecialFeaturesFromJSONTyped(json, false);
}

export function SpecialFeaturesFromJSONTyped(json: any, ignoreDiscriminator: boolean): SpecialFeatures {
    return json as SpecialFeatures;
}

export function SpecialFeaturesToJSON(value?: SpecialFeatures | null): any {
    return value as any;
}

export function SpecialFeaturesToJSONTyped(value: any, ignoreDiscriminator: boolean): SpecialFeatures {
    return value as SpecialFeatures;
}